+++ packages/clients/diplan/example/config.js
@@ -98,6 +98,11 @@ export default {
   },
   draw: {
     enableOptions: true,
+    lassos: [
+      {
+        id: flurstuecke,
+      },
+    ],
     measureOptions: {
       metres: true,
       kilometres: true,
@@ -111,6 +116,7 @@ export default {
         family: 'Arial',
       },
     },
+    toastAction: 'plugin/toast/addToast',
     style: {
       fill: { color: 'rgb(51 117 212 / 50%)' },
       stroke: {