+++ packages/plugins/Draw/src/locales.ts
@@ -40,6 +40,10 @@ export const resourcesDe = {
       label: {
         textSize: 'Textgröße (px) wählen:',
       },
+      lasso: {
+        notInZoomRange:
+          'Der Layer "$t({{serviceName}})" wurde nicht für die Lasso-Funktion genutzt, da er auf der derzeitigen Zoomstufe nicht aktivierbar ist. Bitte zoomen Sie weiter rein.',
+      },
     },
   },
 } as const
@@ -83,6 +87,10 @@ export const resourcesEn = {
       },
       label: {
         textSize: 'Choose text size (px):',
+      },
+      lasso: {
+        notInZoomRange:
+          'The layer "$t({{serviceName}})" was not used for the lasso as it is not activatable on the current zoom level. Please zoom in further.',
       },
     },
   },